Fixed issue with blending using wrong color32 overload

+++ Source/SoftwareRendering/PixelData.cs
@@ -156,10 +156,10 @@ public class PixelData
         var dest = PerformBlending( _pixelDataArray[index], pixel, _pixelDataArray[index], destBlendFactor);
 
         _pixelDataArray[index] = new Color32(
-                source.R + dest.R,
-                source.G + dest.G,
-                source.B + dest.B,
-                source.A + dest.A);
+                (byte)(source.R + dest.R),
+                (byte)(source.G + dest.G),
+                (byte)(source.B + dest.B),
+                (byte)(source.A + dest.A));
     }
 
     /// <summary>
@@ -402,59 +402,59 @@ public class PixelData
             
             case PixelBlendFactor.SourceColor:
                 return new Color32(
-                    factorTargetColor.R * (sourceColor.R / 255.0f),
-                    factorTargetColor.G * (sourceColor.G / 255.0f),
-                    factorTargetColor.B * (sourceColor.B / 255.0f),
-                    factorTargetColor.A * (sourceColor.A / 255.0f));
+                    (byte)(factorTargetColor.R * (sourceColor.R / 255.0f)),
+                    (byte)(factorTargetColor.G * (sourceColor.G / 255.0f)),
+                    (byte)(factorTargetColor.B * (sourceColor.B / 255.0f)),
+                    (byte)(factorTargetColor.A * (sourceColor.A / 255.0f)));
             
             case PixelBlendFactor.OneMinusSourceColor:
                 return new Color32(
-                    factorTargetColor.R * ((255 - sourceColor.R) / 255.0f),
-                    factorTargetColor.G * ((255 - sourceColor.G) / 255.0f),
-                    factorTargetColor.B * ((255 - sourceColor.B) / 255.0f),
-                    factorTargetColor.A * ((255 - sourceColor.A) / 255.0f));
+                    (byte)(factorTargetColor.R * ((255 - sourceColor.R) / 255.0f)),
+                    (byte)(factorTargetColor.G * ((255 - sourceColor.G) / 255.0f)),
+                    (byte)(factorTargetColor.B * ((255 - sourceColor.B) / 255.0f)),
+                    (byte)(factorTargetColor.A * ((255 - sourceColor.A) / 255.0f)));
             
             case PixelBlendFactor.SourceAlpha:
                 return new Color32(
-                    factorTargetColor.R * (sourceColor.A / 255.0f),
-                    factorTargetColor.G * (sourceColor.A / 255.0f),
-                    factorTargetColor.B * (sourceColor.A / 255.0f),
-                    factorTargetColor.A * (sourceColor.A / 255.0f));
+                    (byte)(factorTargetColor.R * (sourceColor.A / 255.0f)),
+                    (byte)(factorTargetColor.G * (sourceColor.A / 255.0f)),
+                    (byte)(factorTargetColor.B * (sourceColor.A / 255.0f)),
+                    (byte)(factorTargetColor.A * (sourceColor.A / 255.0f)));
             
             case PixelBlendFactor.OneMinusSourceAlpha:
                 return new Color32(
-                    factorTargetColor.R * ((255 - sourceColor.A) / 255.0f),
-                    factorTargetColor.G * ((255 - sourceColor.A) / 255.0f),
-                    factorTargetColor.B * ((255 - sourceColor.A) / 255.0f),
-                    factorTargetColor.A * ((255 - sourceColor.A) / 255.0f));
+                    (byte)(factorTargetColor.R * ((255 - sourceColor.A) / 255.0f)),
+                    (byte)(factorTargetColor.G * ((255 - sourceColor.A) / 255.0f)),
+                    (byte)(factorTargetColor.B * ((255 - sourceColor.A) / 255.0f)),
+                    (byte)(factorTargetColor.A * ((255 - sourceColor.A) / 255.0f)));
             
             case PixelBlendFactor.DestinationColor:
                 return new Color32(
-                    factorTargetColor.R * (destColor.R / 255.0f),
-                    factorTargetColor.G * (destColor.G / 255.0f),
-                    factorTargetColor.B * (destColor.B / 255.0f),
-                    factorTargetColor.A * (destColor.A / 255.0f));
+                    (byte)(factorTargetColor.R * (destColor.R / 255.0f)),
+                    (byte)(factorTargetColor.G * (destColor.G / 255.0f)),
+                    (byte)(factorTargetColor.B * (destColor.B / 255.0f)),
+                    (byte)(factorTargetColor.A * (destColor.A / 255.0f)));
             
             case PixelBlendFactor.OneMinusDestinationColor:
                 return new Color32(
-                    factorTargetColor.R * ((255 - destColor.R) / 255.0f),
-                    factorTargetColor.G * ((255 - destColor.G) / 255.0f),
-                    factorTargetColor.B * ((255 - destColor.B) / 255.0f),
-                    factorTargetColor.A * ((255 - destColor.A) / 255.0f));
+                    (byte)(factorTargetColor.R * ((255 - destColor.R) / 255.0f)),
+                    (byte)(factorTargetColor.G * ((255 - destColor.G) / 255.0f)),
+                    (byte)(factorTargetColor.B * ((255 - destColor.B) / 255.0f)),
+                    (byte)(factorTargetColor.A * ((255 - destColor.A) / 255.0f)));
             
             case PixelBlendFactor.DestinationAlpha:
                 return new Color32(
-                    factorTargetColor.R * (destColor.A / 255.0f),
-                    factorTargetColor.G * (destColor.A / 255.0f),
-                    factorTargetColor.B * (destColor.A / 255.0f),
-                    factorTargetColor.A * (destColor.A / 255.0f));
+                    (byte)(factorTargetColor.R * (destColor.A / 255.0f)),
+                    (byte)(factorTargetColor.G * (destColor.A / 255.0f)),
+                    (byte)(factorTargetColor.B * (destColor.A / 255.0f)),
+                    (byte)(factorTargetColor.A * (destColor.A / 255.0f)));
             
             case PixelBlendFactor.OneMinusDestinationAlpha:
                 return new Color32(
-                    factorTargetColor.R * ((255 - destColor.A) / 255.0f),
-                    factorTargetColor.G * ((255 - destColor.A) / 255.0f),
-                    factorTargetColor.B * ((255 - destColor.A) / 255.0f),
-                    factorTargetColor.A * ((255 - destColor.A) / 255.0f));
+                    (byte)(factorTargetColor.R * ((255 - destColor.A) / 255.0f)),
+                    (byte)(factorTargetColor.G * ((255 - destColor.A) / 255.0f)),
+                    (byte)(factorTargetColor.B * ((255 - destColor.A) / 255.0f)),
+                    (byte)(factorTargetColor.A * ((255 - destColor.A) / 255.0f)));
             
             default:
                 throw new ArgumentOutOfRangeException(nameof(blendFactor), blendFactor, null);
